UrlArchiver now store whole http response.
HTTP::Tiny response are pure hash and can be easily stored, that's cool.

If the response content is binary, a blob is created to store the
content itself. $response->{content} is altered to be a HashRef
in order to reference to the blob.

Text content is stored within the response blob.